Added delete

+++ AdvancedLab/AddressBook/Delete.sh
@@ -5,9 +5,15 @@ read key
 if [ -e contacts_tmp ]; then
 rm contacts_tmp
 fi
-cat contacts |  awk -F, -v k="$key"  '{
+
+cat contacts |  awk -F, -v k="$key" '{
 if(tolower($1) != tolower(k)) {
 					print $0
 				}
 			}' > contacts_tmp
-mv contacts_tmp contacts
+diff --suppress-common-lines -y contacts contacts_tmp
+echo "are you sure to delete above press y to confirm"
+read confirm
+if [ "$confirm" = "y" ]; then
+  mv contacts_tmp contacts
+fi
